VertexPointers component docs

+++ include/vclib/mesh/components/vertex_pointers.cpp
@@ -134,6 +134,12 @@ const Vertex* VertexPointers<Vertex, N, El>::vertexMod(int i) const
 	return Base::container(this).atMod(i);
 }
 
+/**
+ * @brief Sets the i-th vertex of the element.
+ * @param[in] v: The pointer to the vertex to set to the element.
+ * @param[in] i: the position in the container on which set the vertex; the
+ * value must be between 0 and the number of vertices.
+ */
 template<typename Vertex, int N, typename El>
 void VertexPointers<Vertex, N, El>::setVertex(Vertex* v, uint i)
 {
@@ -141,17 +147,35 @@ void VertexPointers<Vertex, N, El>::setVertex(Vertex* v, uint i)
 }
 
 template<typename Vertex, int N, typename El>
-void VertexPointers<Vertex, N, El>::setVertices(const std::vector<Vertex*>& list)
+void VertexPointers<Vertex, N, El>::setVertices(
+	const std::vector<Vertex*>& list)
 {
 	Base::container(this).set(list);
 }
 
+/**
+ * @brief Returns `true` if the container of vertices contains the given vertex,
+ * `false` otherwise.
+ *
+ * @param[in] v: the pointer to the vertex to search.
+ * @return `true` if the container of vertices contains the given vertex,
+ * `false` otherwise.
+ */
 template<typename Vertex, int N, typename El>
 bool VertexPointers<Vertex, N, El>::containsVertex(const Vertex* v) const
 {
 	return Base::container(this).contains(v);
 }
 
+/**
+ * @brief Returns an iterator to the first vertex in the container of this
+ * component that is equal to the given vertex. If no such vertex is
+ * found, past-the-end iterator is returned.
+ *
+ * @param[in] v: the pointer to the vertex to search.
+ * @return an iterator pointing to the first vertex equal to the given
+ * vertex, or end if no such vertex is found.
+ */
 template<typename Vertex, int N, typename El>
 typename VertexPointers<Vertex, N, El>::VertexIterator
 VertexPointers<Vertex, N, El>::findVertex(const Vertex* v)
@@ -159,6 +183,16 @@ VertexPointers<Vertex, N, El>::findVertex(const Vertex* v)
 	return Base::container(this).find(v);
 }
 
+
+/**
+ * @brief Returns a const iterator to the first vertex in the container of this
+ * component that is equal to the given vertex. If no such vertex is
+ * found, past-the-end iterator is returned.
+ *
+ * @param[in] v: the pointer to the vertex to search.
+ * @return a const iterator pointing to the first vertex equal to the given
+ * vertex, or end if no such vertex is found.
+ */
 template<typename Vertex, int N, typename El>
 typename VertexPointers<Vertex, N, El>::ConstVertexIterator
 VertexPointers<Vertex, N, El>::findVertex(const Vertex* v) const
@@ -166,73 +200,149 @@ VertexPointers<Vertex, N, El>::findVertex(const Vertex* v) const
 	return Base::container(this).find(v);
 }
 
+/**
+ * @brief Returns the index of the given vertex in the container of
+ * the element. If the given vertex is not in the container, returns
+ * UINT_NULL.
+ *
+ * @param[in] v: the pointer to the vertex to search.
+ * @return the index of the given vertex, or UINT_NULL if it is not
+ * found.
+ */
 template<typename Vertex, int N, typename El>
 uint VertexPointers<Vertex, N, El>::indexOfVertex(const Vertex* v) const
 {
 	return Base::container(this).indexOf(v);
 }
 
+/**
+ * @brief Returns the index of the given edge composed of the two vertices v1
+ * and v2 in the container of the element.
+ *
+ * The order of the two vertices is not important. The index of the edge
+ * corresponds to the index of the first vertex found in the container. If the
+ * edge formed by the two vertices is not in the container, returns UINT_NULL.
+ *
+ * @param[in] v1: the pointer to the first vertex that compose the edge to
+ * search.
+ * @param[in] v2: the pointer to the second vertex that compose the edge to
+ * search.
+ * @return the index of the edge composed of v1 and v2, or UINT_NULL if it is
+ * not found.
+ */
 template<typename Vertex, int N, typename El>
-uint VertexPointers<Vertex, N, El>::indexOfEdge(const Vertex* v1, const Vertex* v2) const
+uint VertexPointers<Vertex, N, El>::indexOfEdge(
+	const Vertex* v1,
+	const Vertex* v2) const
 {
 	uint vid = indexOfVertex(v1);
 	if (vid == UINT_NULL) {
 		return UINT_NULL;
 	}
-	else if (vertexMod(vid+1) == v2) {
+	else if (vertexMod(vid + 1) == v2) {
 		return vid;
 	}
-	else if (vertexMod((int)vid-1) == v2) {
+	else if (vertexMod((int) vid - 1) == v2) {
 		uint n = vertexNumber();
-		return ((vid-1) % n + n) % n;
+		return (((int) vid - 1) % n + n) % n;
 	}
 	else {
 		return UINT_NULL;
 	}
 }
 
+/**
+ * @brief Resize the container of the vertices to the given size.
+ * @note This function is available only if the container of the Vertices has
+ * dynamic size.
+ * @param[in] n: The new size of the vertices container.
+ */
 template<typename Vertex, int N, typename El>
 void VertexPointers<Vertex, N, El>::resizeVertices(uint n) requires (N < 0)
 {
 	Base::container(this).resize(n);
 }
 
+/**
+ * @brief Pushes in the back of the container the given vertex.
+ * @note This function is available only if the container of the Vertices has
+ * dynamic size.
+ * @param[in] v: The pointer to the vertex to push in the back of the container.
+ */
 template<typename Vertex, int N, typename El>
 void VertexPointers<Vertex, N, El>::pushVertex(Vertex* v) requires (N < 0)
 {
 	Base::container(this).pushBack(v);
 }
 
+/**
+ * @brief Inserts the given vertex in the container at the given position.
+ * @note This function is available only if the container of the Vertices has
+ * dynamic size.
+ * @param[in] i: The position in the container where to insert the vertex.
+ * @param[in] v: The pointer to the vertex to insert in the container.
+ */
 template<typename Vertex, int N, typename El>
-void VertexPointers<Vertex, N, El>::insertVertex(uint i, Vertex* v) requires (N < 0)
+void VertexPointers<Vertex, N, El>::insertVertex(uint i, Vertex* v)
+	requires (N < 0)
 {
 	Base::container(this).insert(i, v);
 }
 
+/**
+ * @brief Removes the vertex at the given position from the container.
+ * @note This function is available only if the container of the Vertices has
+ * dynamic size.
+ * @param[in] i: The position of the vertex to remove from the container.
+ */
 template<typename Vertex, int N, typename El>
 void VertexPointers<Vertex, N, El>::eraseVertex(uint i) requires (N < 0)
 {
 	Base::container(this).erase(i);
 }
 
+/**
+ * @brief Clears the container of vertices, making it empty.
+ * @note This function is available only if the container of the Vertices has
+ * dynamic size.
+ */
 template<typename Vertex, int N, typename El>
 void VertexPointers<Vertex, N, El>::clearVertices() requires (N < 0)
 {
 	Base::container(this).clear();
 }
 
+/**
+ * @brief Returns an iterator to the first vertex in the container of this
+ * component.
+ *
+ * @return an iterator pointing to the begin of this container.
+ */
 template<typename Vertex, int N, typename El>
-typename VertexPointers<Vertex, N, El>::VertexIterator VertexPointers<Vertex, N, El>::vertexBegin()
+typename VertexPointers<Vertex, N, El>::VertexIterator
+VertexPointers<Vertex, N, El>::vertexBegin()
 {
 	return Base::container(this).begin();
 }
 
+/**
+ * @brief Returns an iterator to the end of the container of this component.
+ *
+ * @return an iterator pointing to the end of this container.
+ */
 template<typename Vertex, int N, typename El>
-typename VertexPointers<Vertex, N, El>::VertexIterator VertexPointers<Vertex, N, El>::vertexEnd()
+typename VertexPointers<Vertex, N, El>::VertexIterator
+VertexPointers<Vertex, N, El>::vertexEnd()
 {
 	return Base::container(this).end();
 }
 
+/**
+ * @brief Returns a const iterator to the first vertex in the container of this
+ * component.
+ *
+ * @return an iterator pointing to the begin of this container.
+ */
 template<typename Vertex, int N, typename El>
 typename VertexPointers<Vertex, N, El>::ConstVertexIterator
 VertexPointers<Vertex, N, El>::vertexBegin() const
@@ -240,6 +350,12 @@ VertexPointers<Vertex, N, El>::vertexBegin() const
 	return Base::container(this).begin();
 }
 
+/**
+ * @brief Returns a const iterator to the end of the container of this
+ * component.
+ *
+ * @return an iterator pointing to the end of this container.
+ */
 template<typename Vertex, int N, typename El>
 typename VertexPointers<Vertex, N, El>::ConstVertexIterator
 VertexPointers<Vertex, N, El>::vertexEnd() const
@@ -247,12 +363,44 @@ VertexPointers<Vertex, N, El>::vertexEnd() const
 	return Base::container(this).end();
 }
 
+/**
+ * @brief Returns a lightweight view object that stores the begin and end
+ * iterators of the container of vertices of the element. The view
+ * object exposes the iterators trough the `begin()` and `end()` member
+ * functions, and therefore the returned object can be used in range-based for
+ * loops:
+ *
+ * @code{.cpp}
+ * for (auto* vertex : el.vertices()) {
+ *     // Do something with vertex...
+ * }
+ * @endcode
+ *
+ * @return a lightweight view object that can be used in range-based for loops
+ * to iterate over vertices.
+ */
 template<typename Vertex, int N, typename El>
 auto VertexPointers<Vertex, N, El>::vertices()
 {
 	return View(vertexBegin(), vertexEnd());
 }
 
+/**
+ * @brief Returns a lightweight const view object that stores the begin and end
+ * iterators of the container of vertices of the element. The view
+ * object exposes the iterators trough the `begin()` and `end()` member
+ * functions, and therefore the returned object can be used in range-based for
+ * loops:
+ *
+ * @code{.cpp}
+ * for (const auto* vertex : el.vertices()) {
+ *     // Do something read-only with vertex...
+ * }
+ * @endcode
+ *
+ * @return a lightweight view object that can be used in range-based for loops
+ * to iterate over vertices.
+ */
 template<typename Vertex, int N, typename El>
 auto VertexPointers<Vertex, N, El>::vertices() const
 {
@@ -278,14 +426,16 @@ void VertexPointers<Vertex, N, El>::importPointersFrom(
 			if constexpr (N == Element::VERTEX_NUMBER) {
 				importPtrsFrom(e, base, ebase);
 			}
-			// from polygonal to fixed size, but the polygon size == the fixed face size
+			// from polygonal to fixed size, but the polygon size == the fixed
+			// face size
 			else if constexpr (Element::VERTEX_NUMBER < 0){
 				if (e.vertexNumber() == N) {
 					importPtrsFrom(e, base, ebase);
 				}
 			}
 			else {
-				// do not import in this case: cannot import from a face of different size
+				// do not import in this case: cannot import from a face of
+				// different size
 			}
 		}
 		else {
